pyproject.toml
[project]
name = "prop-recal"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.0",
  "matplotlib>=3.8",
  "pyyaml>=6.0",
  "numpy>=2.4.0",
  "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
